Updated in database queries.

+++ src/main/java/com/example/licious/inventoryDemo/repository/InventoryTransactionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InventoryTransactionRepository extends JpaRepository<InventoryTransactions, Long> {
+public interface InventoryTransactionRepository extends JpaRepository<InventoryTransactions, Integer> {
 }
